project-02-a2a-code-review-network/pyproject.toml
[project]
name = "a2a-code-review-network"
version = "0.1.0"
description = "Multi-framework code review agents networked via the A2A protocol"
requires-python = ">=3.11"
dependencies = [
    # A2A protocol
    "a2a-sdk>=0.3.0",
    # Agent frameworks
    "langgraph>=0.4.0",
    "langchain-anthropic>=0.3.0",
    "crewai>=0.80.0",
    "crewai-tools>=0.30.0",
    "anthropic>=0.40.0",
    "langchain-openai>=0.3.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    # GitHub integration
    "PyGithub>=2.3.0",
    # Config
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
